+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Maybe:
 
 - rename doe, write_does and load_does to `sweep` module `read_sweep`, `write_sweep` ...
 - Route and Routes are pydantic.BaseModel instead of dataclasses
+- composed functions get a unique name. You can compose functions with `toolz.compose`
 
 ## 3.6.6